Added implementation of setting category in item, included category when returning item queries

+++ src/Modules/Inventofree.Module.Item.Core/Command/Category/UpdateCategory/UpdateCategoryCommand.cs
@@ -4,9 +4,9 @@ namespace Inventofree.Module.Item.Core.Command.Category.UpdateCategory
 {
     public class UpdateCategoryCommand : IRequest<bool>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int UpdatedBy { get; set; }
+        public long UpdatedBy { get; set; }
     }
 }
