+++ src/Modules/Inventofree.Module.Item.Core/Command/Item/UpdateItem/UpdateItemCommand.cs
@@ -4,9 +4,9 @@ namespace Inventofree.Module.Item.Core.Command.Item.UpdateItem
 {
     public class UpdateItemCommand : IRequest<bool>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Detail { get; set; }
-        public int UpdatedBy { get; set; }
+        public long UpdatedBy { get; set; }
     }
 }
